refactor(graphviz): add default methods to EdgeToGraphviz trait
- Add HasBranchLength supertrait bound to EdgeToGraphviz
- Provide default implementations for to_graphviz_label() and to_graphviz_weight()
- Replace four identical impl blocks with empty impls (EdgeTimetree, EdgeAncestral, EdgeClock, TestEdge)
- Remove unused format_weight and NwkWriteOptions imports from consumer files
